refactor: migrate branch-changed imports to path aliases
Converts 2+ level relative imports to @/ and @test/ path aliases across
all files modified on this branch, per project-conventions.md rules.

- 8 test files (debate selectors, verifiers, operations) updated to use
  @/debate, @/errors, @/operations, @/config, @test/helpers
- 6 src/context/engine + rules files updated to use @/logger, @/errors,
  @/prd, @/config, @/test-runners, @/utils/errors
- src/config/index.ts: export DebateConfigSchema to satisfy barrel rule
- Baseline updated: 3195 deep-relative imports (↓21 vs prior baseline)
