+++ src/config/index.ts
@@ -42,6 +42,7 @@ export {
   AcceptanceConfigSchema,
 } from "./schema";
 export { ConfiguredModelSchema, ModelTierSchema } from "./schemas-model";
+export { DebateConfigSchema } from "./schemas-debate";
 export { loadConfig, loadConfigForWorkdir, findProjectDir, globalConfigPath } from "./loader";
 export { validateConfig, type ValidationResult } from "./validate"; // @deprecated: Use NaxConfigSchema.safeParse() instead
 export { validateDirectory, validateFilePath, isWithinDirectory, MAX_DIRECTORY_DEPTH } from "./path-security";
