+++ src/context/engine/types.ts
@@ -454,14 +454,14 @@ export interface ContextRequest {
    * Providers that need to know "is this a test file?" or "where is the sibling test?"
    * consult this field instead of hardcoding extensions or directory names.
    */
-  resolvedTestPatterns?: import("../../test-runners/resolver").ResolvedTestPatterns;
+  resolvedTestPatterns?: import("@/test-runners").ResolvedTestPatterns;
   /**
    * Pre-built naxIgnore index for this run.
    * When present, CodeNeighborProvider passes the per-package matchers to the
    * glob dep so user-defined .naxignore patterns suppress files from the
    * reverse-dep scan.
    */
-  naxIgnoreIndex?: import("../../utils/path-filters").NaxIgnoreIndex;
+  naxIgnoreIndex?: import("@/utils/path-filters").NaxIgnoreIndex;
 }
 
 // ─────────────────────────────────────────────────────────────────────────────
